Add kmod-uleds
This is the uleds driver from the 4.10 linux kernel.  It allows
a userspace app to register LED class drivers with the kernel such
that sysfs nodes are created for control.  Brightness change events
are sent to the userspace app to allow it to take whetever action
is necessary to manage the corresponding LED

SW-2198