minim-openwifi-scripts: Add led management scripts
This commit adds the long running update_leds.sh script for detecting
different led states, and the leds.sh script that it calls to actually
change the leds.  The majority of leds.sh is device agnostic, but there
are a few places were some device specific tweaks are made based on the
running board name.

SW-2199