unum: Always build the latest on master
We don't just set PKG_SOURCE_VERSION to HEAD, because that isn't
useful when trying to determine what version of the code was used for
this build.  So use git ls-remote to get the actual hash of master.
Skip the tarball hash check as it will always be changing.  Also, stop
setting PKG_VERSION to v2, which will cause the package version to be
the commit hash, which is needed for the jenkins pipeline script to be
able to determine what commit of the unum-sdk repo to tag.

SW-3936